APP/requirements.txt
aiohappyeyeballs          2.6.1
aiohttp                   3.11.14
aiosignal                 1.3.2
annotated-types           0.7.0
anyio                     4.9.0
async-timeout             4.0.3
attrs                     25.3.0
azureopenai               0.0.1
certifi                   2025.1.31
charset-normalizer        3.4.1
colorama                  0.4.6
contourpy                 1.3.1
cycler                    0.12.1
dataclasses-json          0.6.7
distro                    1.9.0
embeddings                0.0.8
exceptiongroup            1.2.2
faiss-cpu                 1.10.0
fonttools                 4.57.0
frozenlist                1.5.0
greenlet                  3.1.1
h11                       0.14.0
httpcore                  1.0.7
httpx                     0.28.1
httpx-sse                 0.4.0
idna                      3.10
jiter                     0.9.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
kiwisolver                1.4.8
langchain                 0.3.23
langchain-community       0.3.21
langchain-core            0.3.51
langchain-openai          0.3.12
langchain-text-splitters  0.3.8
langsmith                 0.3.30
marshmallow               3.26.1
matplotlib                3.10.1
multidict                 6.2.0
mypy-extensions           1.0.0
numpy                     2.2.4
openai                    1.73.0
orjson                    3.10.16
packaging                 24.2
pillow                    11.1.0
pip                       25.0.1
propcache                 0.3.1
pydantic                  2.11.1
pydantic_core             2.33.0
pydantic-settings         2.8.1
pyparsing                 3.2.3
python-dateutil           2.9.0.post0
python-dotenv             1.1.0
PyYAML                    6.0.2
referencing               0.36.2
regex                     2024.11.6
requests                  2.32.3
requests-toolbelt         1.0.0
rpds-py                   0.24.0
scipy                     1.15.2
setuptools                57.4.0
six                       1.17.0
sniffio                   1.3.1
spicy                     0.16.0
SQLAlchemy                2.0.40
tenacity                  9.1.2
tiktoken                  0.9.0
tqdm                      4.67.1
typing_extensions         4.13.0
typing-inspect            0.9.0
typing-inspection         0.4.0
urllib3                   2.3.0
yarl                      1.18.3
zstandard                 0.23.0